slack-summariser/slack_summary/requirements.txt
boto3
slack_sdk
haystack-ai>=2.15
amazon-bedrock-haystack>=3.8.0
python-dotenv